requirements.txt
streamlit
google-generativeai
PyPDF2
streamlit-lottie
